+++ src/components/schedule/index.ts
@@ -5,4 +5,4 @@ export * from "./schedule-detail-content";
 export * from "./schedule-detail-map";
 export * from "./schedule-add-button";
 export * from "./schedule-modal";
-export * from "./schedule-date-time-modal";
+export * from "./schedule-modal-date-time";
